+++ controller/profileController.php
@@ -12,41 +12,52 @@ function profilePage($userId)
     include 'view/frontend/pages/profile.php';
 }
 /* ***************** MODIFY PROFILE ***************/
-function modifyProfile($userId, $avatar, $first_name, $name, $email, $description)
+function modifyProfile($userId, $avatar, $first_name, $name, $email, $description, $csrfProfileToken)
 {
 
     $userManager = new \Philippe\Blog\Model\UserManager();
 
-    if (!empty($_POST['email'])) {
-        // Testons si le fichier a bien été envoyé et s'il n'y a pas d'erreur
-        if (isset($_FILES['avatar']) AND $_FILES['avatar']['error'] == 0) {
-            // Testons si le fichier n'est pas trop gros
-            if ($_FILES['avatar']['size'] <= 1000000) { // taille du fichier envoyé
-                // Testons si l'extension est autorisée
-                $infosfichier = pathinfo($_FILES['avatar']['name']); // nom du fichier envoyé par le visiteur
-                $extension_upload = $infosfichier['extension']; // on récupère l'extension du fichier
-                $extensions_autorisees = array('jpg', 'jpeg', 'gif', 'png');
-                // on vérifie si l'extension de notre fichier correspond bien à celles qu'on autorise.
-                if (in_array($extension_upload, $extensions_autorisees)) {
-                    // On peut valider le fichier et le stocker définitivement
-                    move_uploaded_file($_FILES['avatar']['tmp_name'], 'public/images/avatar/' . basename($_FILES['avatar']['name'])); // déplace notre fichier de l'emplacement temporaire vers notre serveur
-                    echo "L'envoi a bien été effectué !";
+    if (!empty($_POST['email'])) 
+    {
+        if (isset($_SESSION['csrfProfileToken']) AND isset($csrfProfileToken) AND !empty($_SESSION['csrfProfileToken']) AND !empty($csrfProfileToken)) 
+        {
+            if ($_SESSION['csrfProfileToken'] == $csrfProfileToken) 
+            {
+                if (isset($_FILES['avatar']) AND $_FILES['avatar']['error'] == 0) 
+                {
+            
+                    if ($_FILES['avatar']['size'] <= 1000000) { 
+                
+                        $infosfichier = pathinfo($_FILES['avatar']['name']); 
+                        $extension_upload = $infosfichier['extension']; 
+                        $extensions_autorisees = array('jpg', 'jpeg', 'gif', 'png');
+                        
+                        if (in_array($extension_upload, $extensions_autorisees)) {
+                            
+                            move_uploaded_file($_FILES['avatar']['tmp_name'], 'public/images/avatar/' . basename($_FILES['avatar']['name'])); 
+                            echo "L'envoi a bien été effectué !";
+                        }
+                    }
+                }
+                $modifiedProfile = $userManager->modifyProfileRequest($userId, $avatar, $first_name, $name, $email, $description);
+                if ($modifiedProfile === false) 
+                {
+                    throw new Exception('Impossible de modifier le profil');
+                }
+                else 
+                {
+                    $_SESSION['flash']['success'] = 'Modification effectuée !';
+                    profilePage($_SESSION['id']);
+                    exit();
+                    unset($_SESSION['avatar']);
+                    $_SESSION['avatar'] = $avatar;
                 }
             }
+            else
+            {
+                echo "Erreur de vérification";
+            }
         }
-        $modifiedProfile = $userManager->modifyProfileRequest($userId, $avatar, $first_name, $name, $email, $description);
-        if ($modifiedProfile === false) {
-            throw new Exception('Impossible de modifier le profil');
-        }
-        else 
-        {
-            $_SESSION['flash']['success'] = 'Modification effectuée !';
-            profilePage($_SESSION['id']);
-            exit();
-            unset($_SESSION['avatar']);
-            $_SESSION['avatar'] = $avatar;
-        }
-            
     }
     else 
     {
@@ -56,18 +67,42 @@ function modifyProfile($userId, $avatar, $first_name, $name, $email, $descriptio
     }
 }
 /* ***************** DELETE ACCOUNT ***************/
-function deleteAccount($userId)
+function deleteAccount($userId, $csrfDeleteAccountToken)
 {
     $userManager = new \Philippe\Blog\Model\UserManager();
     $session = new Session();
-    $deleteAccount = $userManager->deleteAccountRequest($userId);
-    $session->stopSession();
+    
+    if (isset($userId)) 
+    {
+        if (isset($_SESSION['csrfDeleteAccountToken']) AND isset($csrfDeleteAccountToken) AND !empty($_SESSION['csrfDeleteAccountToken']) AND !empty($csrfDeleteAccountToken)) 
+        {
+            if ($_SESSION['csrfDeleteAccountToken'] == $csrfDeleteAccountToken) 
+            {
+                $deleteAccount = $userManager->deleteAccountRequest($userId);
+                $session->stopSession();
 
-    if ($deleteAccount === false) {
-        throw new Exception('Impossible de supprimer le profil');
+                if ($deleteAccount === false) 
+                {
+                    throw new Exception('Impossible de supprimer le profil');
+                }
+                else 
+                {
+                    //header('Location: index.php?action=blog');
+                    home();
+                    exit();
+                }
+            }
+            else
+            {
+                echo "Erreur de vérification";
+            }
+        }
     }
-    else {
-        header('Location: index.php?action=blog');
+    else 
+    {
+        $_SESSION['flash']['danger'] = 'Aucun id ne correspond à cet utilisateur !';
+        profilePage();
+        exit();
     }
 }
 /* ***************** PUBLIC PROFILE ***************/
